requirements.txt
azure-core==1.30.1
azure-identity==1.15.0
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
idna==3.6
isodate==0.6.1
msal==1.28.0
msal-extensions==1.1.0
numpy==1.26.4
pandas==2.2.1
portalocker==2.8.2
pyarrow==15.0.2
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1